requirements_diffusion.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.24.0
Pillow>=9.5.0
numpy>=1.24.0
opencv-python>=4.8.0
# Optional heavy dependencies (commented out for faster installation)
# transformers>=4.35.0
# accelerate>=0.24.0  
# safetensors>=0.4.0
# xformers>=0.0.22